OnlineLearning-Alex/requirements.txt
Theano==0.6.0
nose==1.3.3
numpy==1.8.1
pypng==0.0.17
scipy==0.14.0
